+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- API method `set_payload()` now sets only the given key within payload table
+
 ## [2.0.0] - 2018-04-03
 ### Changed
 - Rename API method: `quit()` -> `leave()`